Add ability for admin to edit comments
Add editComment method to comment service. Add edit comment action to admin
comment servlet which calls the service. Add field for "edited" in comment
resource and pass that data between client and server sides. Add note on comment
view for edited comments. Add methods to clientside comment service to
communicate with server. Break out similar functions into helper methods between
edit and delete functionality and share the comments modal.

TODO: refactor comment view to use service.
TODO: refactor comment servlet to use service and AJAX.

+++ ui/src/main/resources/jcr_root/etc/clientlibs/admin/js/commentModalController.js
@@ -3,14 +3,23 @@
  * actions such as deletion, marking as spam, and marking as ham. The author can
  * also edit comments.
  */
-app.controller('CommentModalController', function ($scope, $modalInstance, CommentService, comment) {
+app.controller('CommentModalController', function ($scope, $modalInstance, CommentService, action, comment) {
 
-  $scope.comment = comment;
+  $scope.editMode = action === 'edit';
+  $scope.deleteMode = action === 'delete';
+  $scope.comment = angular.copy(comment);
 
   $scope.ok = function () {
-    CommentService.deleteComment($scope.comment).success(function(data){
-      $modalInstance.close({success: true, comment: comment});
-    });
+    if ($scope.deleteMode) {
+      CommentService.deleteComment($scope.comment).success(function(data){
+        $modalInstance.close({success: true, comment: comment});
+      });
+    } else if ($scope.editMode) {
+      CommentService.editComment($scope.comment).success(function(data){
+        $scope.comment.edited = true;
+        $modalInstance.close({success: true, comment: $scope.comment});
+      });
+    }
   };
 
   $scope.cancel = function () {
